Bugfix: Scene Transition on Start
Fixed a bug where the scene would automatically transition back to the main menu on start. When I did some poking around I found that the pit was detecting part of the tilemap as a body, and upon any detection with a body, it would restart the scene. Now, the pit makes sure that the body it detects is the player before it transitions the scene.